+++ src/space/vistar/embrowser/Browser/BrowserConfig.java
@@ -1,0 +1,5 @@
+package space.vistar.embrowser.Browser;
+
+public class BrowserConfig {
+    public static String userAgent = "Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:60.0) Gecko/20100101 Firefox/61.0";
+}